Make code generation know its target files
First, rename 'generate_expand_function.py' to 'autogenerate.py'. It's a
much shorter name. Moreover, it's supposed to be the only one file for
code generation code: if new generated function is needed at some moment
in the future, its generation code must be simply added to
'autogenerate.py' script as a function.

Next, old script used to print generated code to STDOUT, not caring about
where it's supposed to be written. Now it writes its result to a file by
itself. Currently it generated only one file:

   (DIR)/src/rvlm/paraform/autogen.py

where (DIR) is a directory 'autogenerate.py' script resides in. This name
is very unlikely to change, but don't forget that it's hardcoded now.

Finally, make a pair of minor improvements:

 - rename 'gencode_for_expand_function' to 'generate_expand_function_code'
 - add some 'noinspection' comments to make PyCharm silent
 - replace fictional 'Error' with 'ValueError'

Signed-off-by: Pavel Kretov <firegurafiku@gmail.com>